feat(scratch): factspike -- piece-embedding transfer preserves a taught fact
Tests whether a scratch slot bound directly to a word's OWN already-trained
vocab-piece embedding(s) (not byte-decomposed like every prior curriculum)
still carries a factual association learned via ordinary text training, with
zero textual restatement in the eval prompt. No engine change needed:
ScratchTable::bind() stores ids verbatim and encode_slot() never cared
whether bound ids were byte codes or real piece ids.

Real result on a dedicated forced-multi-piece d96 toy model: baseline (fact
learnable via text) peaks at 1.00; scratch (piece-transfer, no text) peaks
at 0.56 vs 0.125 chance, repeated across many rounds; held-out negative
control's own peak (0.33, n=3) is consistent with small-n noise rather than
leakage. Full trajectories and reasoning in docs/FACTSPIKE.md.

Also fixes a real MSVC 19.51 regression found while standing up the toy
model's build: #pragma omp for requires a signed loop variable.

+++ src/backend_cpu.cpp
@@ -1372,12 +1372,15 @@ float train_batch(const int* data, const std::size_t* starts, int batch, int T,
         }
         // (implicit barrier above: every thread's grad slot is complete)
         const int nthreads = omp_get_num_threads();
+        // OpenMP's `for` worksharing construct requires a SIGNED loop variable (unlike `simd`, which
+        // MSVC accepts size_t for, just ignoring reduction clauses on it) -- ptrdiff_t matches size_t's
+        // width so PARAM_FLOATS (a size_t) still fits, just signed to satisfy the spec.
         #pragma omp for schedule(static)
-        for (std::size_t i = 0; i < PARAM_FLOATS; ++i) {
+        for (std::ptrdiff_t i = 0; i < static_cast<std::ptrdiff_t>(PARAM_FLOATS); ++i) {
             float s = 0.f;
             for (int t = 0; t < nthreads; ++t)
-                s += g_workers[t]->grad[i];
-            g_param_grad[i] = s;
+                s += g_workers[t]->grad[static_cast<std::size_t>(i)];
+            g_param_grad[static_cast<std::size_t>(i)] = s;
         }
     }
     return static_cast<float>(total / batch);
